Rewrite to add nixvim and be more modular
The big changes are to add nixvim to manage neovim and moving to a more
module centric config. This is in support of moving my work user to a
repo at work due to wanting more worspecific settings. The hope is that
any thing generic goes here in the upstream config and specifics are
done in another repo. This also removes the work configs and previous
neovim setups.