Remove dependence on self in favor of assertions
previously the modules in `baseline` used the self passed in to reference
packages and overlays but this break compatibility for down stream
users. Instead the user needs to apply the overlay that is provided by
the flake. This is enfoced with assertions to improve the legibility of
the errors.

There is also some reorganization and cleanup in the `gpu` module to
improve readability and performance by not running the nvidia eval if
the nvidia override is not set. `self` is still used but only in the
case of the impure eval that is needed by nvidia.