Reshape to be more reusable
The current structure indicates that all things are meant to be modular
and reusable. While the pieces were idempotent they are also super
opinionated. So I created the `config` folder that will hold all
configuration though still idempotent because that is the way I prefer.
`modules` will now be truly reusable components that are intended for
others to configure themselves. i.e. they do not take an opinion after
just setting an enable flag. Instead, they can be used to expand a
configuration.